replace f32 with f64

+++ src/Neuron.go
@@ -1,15 +1,15 @@
 package main
 
 type NeuronProcessor interface {
-	ProcessNeuron(microbe Microbe) float32
+	ProcessNeuron(microbe Microbe) float64
 }
 
 type Neuron struct {
 	inputs []Synapse
 }
 
-func (neuron *Neuron) GetSynapticSum(microbe Microbe) float32 {
-	var value float32 = 0
+func (neuron *Neuron) GetSynapticSum(microbe Microbe) float64 {
+	var value float64 = 0
 	for _, synapse := range neuron.inputs {
 		value += synapse.GetValue(microbe)
 	}
